Hook up dotnet-transform-xdt to publish.
When publishing, the `_WebConfigTransform` target in Microsoft.NET.Sdk.Publish.TransformFiles.targets runs, and it adds the system.webServer/handlers/aspNetCore
handler. This defines a `RemoveHandlersFromWebConfig` target that runs *after* `WebConfigTransform` and applies `Web.RemoveHandlers.config` to remove the handler. The end result is that the `aspNetCore` handler is removed from the published Web.config.